requirements.txt
fastapi
uvicorn[standard]
deepface
opencv-python-headless
retina-face
onnxruntime
numpy
Pillow
